pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dsat"
version = "0.1.0"
description = "Data Science Assistant Tool - Kedro Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kedro>=1.0.0",
    "kedro-datasets[pandas]>=3.0.0",
    "pandas>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-aiplatform>=1.0.0",
    "vertexai>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
dsat = "dsat.__main__:main"

[tool.kedro]
package_name = "dsat"
project_name = "DSAT"
kedro_init_version = "1.1.1"
tools = "none"
example_pipeline = "false"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"

[tool.kedro_telemetry]
project_id = "c2040afb3f474dcfbe8499645fedaa83"
